Fixes: additional modifications for handling "floats" with Java doubles

+++ pika/src/tokens/FloatingToken.java
@@ -8,7 +8,7 @@ public class FloatingToken extends TokenImp {
 	protected FloatingToken(TextLocation location, String lexeme) {
 		super(location, lexeme);
 	}
-	protected void setValue(float value) {
+	protected void setValue(double value) {
 		this.value = value;
 	}
 	public double getValue() {
@@ -17,7 +17,7 @@ public class FloatingToken extends TokenImp {
 	
 	public static FloatingToken make(TextLocation location, String lexeme) {
 		FloatingToken result = new FloatingToken(location, lexeme);
-		result.setValue(Float.parseFloat(lexeme));
+		result.setValue(Double.parseDouble(lexeme));
 		return result;
 	}
